Fix wrong implementation of ISTransient for value type IDs.
Closes #8
