+++ tests/Fluxera.Entity.UnitTests/EntityWithGuidKeyTransientTests.cs
@@ -1,0 +1,38 @@
+﻿namespace Fluxera.Entity.UnitTests
+{
+	using System;
+	using FluentAssertions;
+	using Fluxera.Entity.UnitTests.EmployeeAggregate;
+	using NUnit.Framework;
+
+	[TestFixture]
+	public class EntityWithGuidKeyTransientTests
+	{
+		[Test]
+		public void ShouldReturnTrueForTransientEntityForDefaultValue()
+		{
+			EmployeeGuid employee = new EmployeeGuid
+			{
+				Name = "James Bond",
+				EmployeeNumber = 8051007,
+				Salary = 1_000_000,
+			};
+
+			employee.IsTransient.Should().BeTrue();
+		}
+
+		[Test]
+		public void ShouldReturnTrueForTransientEntityForEmpty()
+		{
+			EmployeeGuid employee = new EmployeeGuid
+			{
+				ID = Guid.Empty,
+				Name = "James Bond",
+				EmployeeNumber = 8051007,
+				Salary = 1_000_000,
+			};
+
+			employee.IsTransient.Should().BeTrue();
+		}
+	}
+}
